Moved didUpdateLocations delegate method to a new class called LocationController Created QiblaDirectionProtocol to communicate with the View class Moved getBearing method from LogicController to Model
